requirements.txt
asgiref==3.7
Django==4.2
djangorestframework==3.14
djangorestframework_simplejwt==5.2
django-filter==23.2
drf-spectacular==0.26.3
PyJWT==2.6
sqlparse==0.4
typing_extensions==4.8
certifi==2023.7.22
charset-normalizer==3.2
coreapi==2.3.3
coreschema==0.0.4
idna==3.4
inflection==0.5
itypes==1.2
Jinja2==3.1
MarkupSafe==2.1
packaging==23.1
pyparsing==3.1
psycopg2==2.9.8
pytz==2023.3
requests==2.31
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
uritemplate==4.1
urllib3==2.0
pytest-cov==4.1
pytest-django==4.5